+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.13.0 - Upcoming
+
+### Added
+
+- Add locale-independent Utils::asciiToLower and Utils::asciiToUpper helpers
+
+### Changed
+
+- Use locale-independent ASCII case folding everywhere case is normalized
+
 ## 2.12.5 - Upcoming
 
 ### Changed